# docs/generated/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BNPI PMS API",
    "version": "1.0.0",
    "description": "Welcome to the BNPI PMS API documentation. This REST API powers BNPI PMS backend services.",
    "contact": {
      "name": "BNPI PMS Support",
      "email": "support@bnpi-pms.example",
      "url": "https://bnpi-pms.example/support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "termsOfService": "https://bnpi-pms.example/terms"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Development server"
    },
    {
      "url": "https://api.bnpi-pms.example/api",
      "description": "Production server"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key for authentication"
      }
    },
...
  },
  "paths": {
...
    "/api/order/{id}": {
      "get": {
        "summary": "Get order by ID",
        "description": "Retrieve a specific order by its unique identifier with optional field selection",
        "tags": [
          "Order"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{24}$"
            },
            "description": "Order ID (MongoDB ObjectId format)",
            "example": "507f1f77bcf86cd799439011"
          },
          {
            "in": "query",
            "name": "fields",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated list of fields to include (supports nested fields with dot notation)",
            "example": "id,name,description,type"
          }
        ],
        "responses": {
          "200": {
            "description": "Order retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "order": {
                              "$ref": "#/components/schemas/Order"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "patch": {
        "summary": "Update order",
        "description": "Update order data by ID (partial update)",
        "tags": [
          "Order"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{24}$"
            },
            "description": "Order ID (MongoDB ObjectId format)",
            "example": "507f1f77bcf86cd799439011"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "minProperties": 1,
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "description": "Order name",
                    "example": "Updated Email Order"
                  },
                  "description": {
                    "type": "string",
                    "description": "Order description",
                    "example": "Updated description for the order"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "email",
                      "sms",
                      "push",
                      "form"
                    ],
                    "description": "Order type for categorization",
                    "example": "email"
                  },
                  "isDeleted": {
                    "type": "boolean",
                    "description": "Soft delete flag",
                    "example": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "order": {
                              "$ref": "#/components/schemas/Order"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "summary": "Delete order",
        "description": "Permanently delete a order by ID",
        "tags": [
          "Order"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{24}$"
            },
            "description": "Order ID (MongoDB ObjectId format)",
            "example": "507f1f77bcf86cd799439011"
          }
        ],
        "responses": {
          "200": {
            "description": "Order deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "description": "Empty object for successful deletion"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/order": {
      "get": {
        "summary": "Get all orders",
        "description": "Retrieve orders with advanced filtering, pagination, sorting, field selection, and optional grouping",
        "tags": [
          "Order"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number for pagination",
            "example": 1
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "description": "Number of records per page",
            "example": 10
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            },
            "description": "Sort order for results",
            "example": "desc"
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Field to sort by or JSON object for multi-field sorting",
            "example": "createdAt"
          },
          {
            "in": "query",
            "name": "fields",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated list of fields to include (supports dot notation)",
            "example": "id,name,description,type"
          },
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Search query to filter by name or description",
            "example": "welcome email"
          },
          {
            "in": "query",
            "name": "filter",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "JSON array of filter objects for advanced filtering",
            "example": "[{\"type\":\"email\"},{\"isDeleted\":false}]"
          },
          {
            "in": "query",
            "name": "groupBy",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Group results by a field name",
            "example": "type"
          },
          {
            "in": "query",
            "name": "document",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "true"
              ]
            },
            "description": "Include order documents in response"
          },
          {
            "in": "query",
            "name": "pagination",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "true"
              ]
            },
            "description": "Include pagination metadata in response"
          },
          {
            "in": "query",
            "name": "count",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "true"
              ]
            },
            "description": "Include total count in response"
          }
        ],
        "responses": {
          "200": {
            "description": "Templates retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "orders": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/Order"
                              },
                              "description": "Present when document=\"true\" and no groupBy"
                            },
                            "groups": {
                              "type": "object",
                              "additionalProperties": {
                                "type": "array",
                                "items": {
                                  "$ref": "#/components/schemas/Order"
                                }
                              },
                              "description": "Present when groupBy is used and document=\"true\""
                            },
                            "count": {
                              "type": "integer",
                              "description": "Present when count=\"true\""
                            },
                            "pagination": {
                              "$ref": "#/components/schemas/Pagination",
                              "description": "Present when pagination=\"true\""
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "summary": "Create new order",
        "description": "Create a new order with the provided data",
        "tags": [
          "Order"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "description": "Order name",
                    "example": "Email Welcome Order"
                  },
                  "description": {
                    "type": "string",
                    "description": "Order description",
                    "example": "Welcome email order for new users"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "email",
                      "sms",
                      "push",
                      "form"
                    ],
                    "description": "Order type for categorization",
                    "example": "email"
                  },
                  "isDeleted": {
                    "type": "boolean",
                    "description": "Soft delete flag",
                    "default": false
                  }
                }
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "description": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "isDeleted": {
                    "type": "boolean"
                  }
                }
              }
            },
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "description": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "isDeleted": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "order": {
                              "$ref": "#/components/schemas/Order"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  }
}